+++ bin/nanocl/src/models/stats.rs
@@ -1,0 +1,6 @@
+use clap::Parser;
+
+#[derive(Default, Parser)]
+pub struct StatsOpts {
+  pub names: Vec<String>,
+}