+++ bin/nanocld/src/services/process/mod.rs
@@ -34,6 +34,7 @@ pub fn ntex_config(config: &mut web::ServiceConfig) {
   config.service(wait_processes);
   config.service(stats_processes);
   config.service(count_processes);
+  config.service(process_stats_by_name);
 }
 
 #[cfg(test)]
@@ -72,7 +73,19 @@ mod tests {
       .await;
     test_status_code!(res.status(), http::StatusCode::OK, "basic cargo stats");
   }
-
+  #[ntex::test]
+  async fn test_process_stats_by_name() {
+    let system = gen_default_test_system().await;
+    let client = system.client;
+    let res = client
+      .send_get("/process/nstore.system.c/stats", None::<String>)
+      .await;
+    test_status_code!(
+      res.status(),
+      http::StatusCode::OK,
+      "process by name stats"
+    );
+  }
   #[ntex::test]
   async fn list_by() {
     let system = gen_default_test_system().await;