+++ crates/nanocld_client/src/process.rs
@@ -213,6 +213,26 @@ impl NanocldClient {
       .await?;
     Ok(Self::res_stream(res).await)
   }
+  /// The stats are streamed as a [Receiver](Receiver) of [stats](Stats)
+  /// It get the stats by key name(returned by nanocl ps) instead of name and kind
+  /// ## Example
+  ///
+  /// ```no_run, ignore
+  /// use nanocld_client::NanocldClient;
+  /// let client = NanocldClient::connect_to("http://localhost:8585", None);
+  /// let stats = client.stats_process_by_name("nstore.system.c");
+  /// ```
+  ///
+  pub async fn stats_process_by_name(
+    &self,
+    name: &str,
+  ) -> HttpClientResult<Receiver<HttpResult<ProcessStats>>> {
+    let query: Option<&ProcessStatsQuery> = None;
+    let res = self
+      .send_get(&format!("/process/{name}/stats"), query)
+      .await?;
+    Ok(Self::res_stream(res).await)
+  }
 
   /// Inspect a process by it's name
   ///
@@ -257,6 +277,20 @@ mod tests {
         "nstore",
         Some(&ProcessLogQuery::of_namespace("system")),
       )
+      .await
+      .unwrap();
+    let _out = rx.next().await.unwrap().unwrap();
+  }
+
+  #[ntex::test]
+  async fn process_stats_by_name() {
+    let client = NanocldClient::connect_to(&ConnectOpts {
+      url: "http://nanocl.internal:8585".into(),
+      ..Default::default()
+    })
+    .expect("Failed to create a nanocl client");
+    let mut rx = client
+      .stats_process_by_name("nstore.system.c")
       .await
       .unwrap();
     let _out = rx.next().await.unwrap().unwrap();